--- packages/react-cosmos-plugin-open-file/src/index.ts
@@ -1,1 +1,0 @@
-export { openFileDevServerPlugin } from './openFile';
