+++ packages/react-cosmos-plugin-open-fixture/.npmignore
