+++ packages/react-cosmos-plugin-open-fixture/index.d.ts
