+++ packages/react-cosmos-plugin-open-fixture/index.js
