+++ packages/react-cosmos-plugin-open-fixture/src/@types/launch-editor.d.ts
