+++ packages/react-cosmos-plugin-open-fixture/src/index.ts
@@ -1,0 +1,1 @@
+export { openFixtureDevServerPlugin } from './openFixture';
